[Rename] SetEventView에 AddScheduleSelectView 병합 [Fix] 이미지 선택 초기화 문제 해결
