+++ timestone/ContentView.swift
@@ -19,7 +19,7 @@ struct ContentView: View {
                     Text("일정")
                 }
             // 맞는 View로 바꿔서 적용하시면 됩니다.
-            AddScheduleView()
+            SetEventView()
                 .tabItem {
                     Image(systemName: "plus.square")
                     Text("일정 추가")
